Changes to Chapter 1, 2 and 3 (and appendix)
Bunch of changes today.
Appendix A: Added an appendix to show how to turn on warnings for Replit. Removed the AP code material
Chapter 1: Added my replit references to the index
Chapter 2: Added replit references to the index, changed the links to the projects
Chapter 3: Changed the C typecast to static cast, changed the compares to better fit on the page, added a mention to problems about the = operator in if statements, Put the print parity function info in the Function chapter. Brought the and and or material into this chapter, added switch statements to this chapter.
Chapter 4: Now has the print parity function info
Chapter 5: removed &&, || and NOT
Chapter 6: Fixed link to print parity function material
Chapter 13: removed the switch code and put in different chapter
contrib: added some notes talking about changes
code: added code files from chapter 1 and 2