some tweaks to the book, mostly reordering
Added rough draft for a couple flow charts. Added different colors for the comments, and strings. Moved the for loop to the iteration chapter. Added the flowchart and some stubs for sections in chapter 3. Moved the iteration chapter to after the selection chapter. Took the function material out of the iteration chapter and put it after the last function chapter.