Added more about switches and standard versions
Added a section in the appendix that mentions how to switch standards in Reddit
Added another flowchart image in chapter 3, added more material on AND and OR, and more about the switch statement.
I also fixed the new figures, so they are not all at the top of the pages.