Started adding the Advanced topics
Added link to HelloParsons in Chapter 2
Fixed a section talking about <, <= and more, because the character did not work in latex
Added a partially finished Advanced Topics chapter (with Lambda Expressions, Functors and Std:: Functions) -- In VERY rough writing. Needs a lot of work.
Chapter 16 will NOT show up in the pdf yet.